+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 147:
 
 * Don't use boost::string_ref
+* Use iterator wrapper in detail::buffers_range
 
 --------------------------------------------------------------------------------